+++ src/App.tsx
@@ -48,6 +48,9 @@ const App: FC = () => {
                 <Route path="/my-radars/*" element={<MyRadars />} />
                 <Route path="/constructor" element={<Constructor />} />
                 <Route path="/single" element={<TechSinglePage />} />
+                <Route path="/constructor" element={<Constructor />}>
+                    <Route path=":versionSlug" element={<Constructor />} />
+                </Route>
                 <Route path="/my-radars/radar-constructor" element={<RadarConstructor />} />
                 <Route path="/admin/*" element={<RequireAuth />} />
             </Routes>
